requirements.txt
transformers
sentence-transformers
faiss-cpu
langchain
langchain-text-splitters
numpy
fastapi
uvicorn[standard]
